Update calculateGasPrice to fetch from config (#1517)
* Update calculateGasPrice to fetch from config

* Remove unused gas estimation function

* Remove unused gasCost estimate function

* Add gasParameter to gasPriceOracle
