--- src/logic/safe/api/index.ts
@@ -1,1 +1,0 @@
-export * from './estimateTxGas'
