+++ src/logic/safe/transactions/index.ts
@@ -1,4 +1,3 @@
-export * from './gas'
 export * from './send'
 export * from './offchainSigner'
 export * from './txHistory'
